+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "watermelon" extension will be documented in this fil
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [2.0.1]
+
+- Fixed Gitlab MR mapping
+- Removed dependency on vscode.git
+- Removed octokit
+
 ## [2.0.0]
 
 - Moved UI to use treeview